facu/tpfinal/elEden2/backend/requirements.txt
# Django and Django REST Framework
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.61.1

# Environment Variables
python-dotenv==1.0.1

# Validation & Filtering
django-filter==23.5
django-rest-auth==0.9.5

# Development & Debug
django-debug-toolbar==4.4.6
factory-boy==3.3.1

# Production
gunicorn==21.2.0
whitenoise==6.7.0

# Testing
pytest==8.3.2
pytest-django==4.8.0
coverage==7.6.1
